Log outgoing server TLS information correctly
We we falsely logging/noticing that the connection was plaintext when it
was not. This was because the TLS connection state was not set when we
connected outgoing to servers.
This also correctly error checks the handshake now. Previously the error
was ignored.
I also restructure how we gather the TLS connection state to be on
demand rather than stored once as storing once proved to be a source of
bugs.