Add a delay between outbound link attempts.
This is to try to avoid race conditions where we connect to the same
network multiple times. Previously we would initiate connection to all
servers at the "same" time. I've seen at least one case where this lead
to noisy and collisions.

This changes the behaviour of the ConnectAttemptTime config option to be
global than per server.

I also correct comments mentioning old names for types (ServerClient,
UserClient).

I also now check validity of channel name in SJOIN. And in JOIN from a
server, we no longer close the link if we see a channel name we don't
like. We just tell the other side.

Finally, add a validity check for the NICK given in server NICK
commands.