Fix bug where remote NICK commands caused collisions incorrectly
The "new" nick given was actually the old nick, so collision logic was
run incorrectly.

This also adds an oper message when there is a collision.

This also checks for whether we found a user in the Users map.
Previously due to the original bug we would panic.

This commit also includes a first change to make us return
targets+messages rather than sending them directly in functions. The
idea is this will make unit testing possible.