Delink from a server that tells us about a server we already know
This could happen if two servers link to the same server at about the
same time.

The impact of not doing this is desynchronization and crashes. Crashes
come from the fact that we'd end up delinking anyway, but our state
would be corrupted. User records could stay in a state where if we try
to send a message along to the user's server (such as a WHOIS), we'd end
up sending a message to the remote server's channel that was cleaned up.